feat: payment methods
When an invoice is created, a `total` price and `currency` are provided
this represents how much has to be payed in order to consider this
invoice as completed. The invoice can be payed
using any of the provided `paymentMethods`

The `exchange-rates` plugin will be used to convert the amount from the
invoice's currency to the paymentMethod's

An invoice's currency may only be used for defining it's value but may not be a
supported paymentMethod. i.e, you can define the price of the invoice
in USD and require payment using XMR and BTC

The address for each paymentMethod is calculated on demand,
when the user requests to pay using that method; with the exception of
the paymentMethod for the invoice's currency.
This one will be generated when the invoice is created.

An invoice may be payed using multiple paymentMethods.
The remaining price will be calculated using the original exchange rate.
i.e, you can pay an 1 XMR invoice with 0.5 XMR and the BTC
equivalent of 0.5 XMR

We can determine when an invoice has been fully payed by representing it's
payments as percentages of the invoice's total amount.